% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaastnr.R
\name{get_weather_data}
\alias{get_weather_data}
\title{Get weather data}
\usage{
get_weather_data(station_number, year, path = NULL)
}
\arguments{
\item{station_number}{factor}

\item{year}{integer}

\item{path}{factor, optional}
}
\value{
data.frame
}
\description{
Loads and cleans weather data for a given NOAA station ID and year.
Returns a dataframe containing a time series of air temperature,
atmospheric pressure, wind speed, and wind direction. Also saves a copy of
the raw data file downloaded from the NOAA FTP server at
'ftp.ncei.noaa.gov/pub/data/noaa/'.
}
\examples{
get_weather_data('911650-22536', 2020)
}
