% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaastnr.R
\name{get_stations_info}
\alias{get_stations_info}
\title{Get stations information}
\usage{
get_stations_info(country = "all")
}
\arguments{
\item{country}{character, optional}
}
\value{
data.frame
}
\description{
Downloads and cleans the data of all stations available at
\url{ftp://ftp.ncei.noaa.gov/pub/data/noaa}.
}
\examples{
get_stations_info(country = "US")
}
